requirements.txt
django
pillow
ruff
psycopg2-binary
dotenv
dj_database_url
gtts
django-storages
python-decouple